# perf(server): optimize mapAsset