Noticed the `mapAsset` function was taking way too long. This seems to be caused by the object spread syntax, after removing it performance got much better.

Tested using `AssetService.getTimeBuckets`, specifically when the `mapAsset(asset, { withStack: true })` case gets returned. It would be useful to know if someone else can replicate these results (only includes calls to mapAsset function, nothing else).

| Assets     | Before   | After |
|------------|----------|-------|
| 100        | 3ms      | <1ms  |
| 1000       | 26ms     | 1ms   |
| 10000      | 397ms    | 6ms   |